==> code/02_cleaning_reddit_data.ipynb <==
# Cleaning and EDA
---

## Contents
---
- [Library Imports](#Library-Imports)
- [Cleaning](#Cleaning)

### Library Imports
---

#Imports
import pandas as pd

### Cleaning
---

# Pull in the corpus.csv and transform into a dataframe - drop the index included in the csv with index_col = [0]
corpus = pd.read_csv('../data/corpus.csv', index_col = [0])

corpus.info()

**The following steps will combine the title and selftext columns into one column, separated with a white space.  The column will be called 'text'.**

# While all posts have text in their titles, about 32% do not have text in the body of the post.  These are likely images.  Before I can concantenate the columns together, I will replace NaNs with a white space.
corpus.isnull().sum()

corpus.fillna(' ', inplace = True)

# No NaNs! \o/
corpus.isnull().sum()

#Concatenate the title and selftext columns together into single column of text.
corpus['text'] = corpus['title'] + ' ' + corpus['selftext']

corpus.head()

# Drop the title and selftext columns as they are now combined into one column.
columns_to_drop = ['title', 'selftext']

corpus.drop(columns = columns_to_drop, inplace = True)

corpus.head()

**For modeling, I will change the 'fp' subreddit classifier to 1 and the 'pens' subreddit classifier to 0.**

corpus.replace({'fp':1, 'pens':0}, inplace = True)

corpus

**Read the cleaned datafram back into a csv file**

corpus.to_csv('../data/cleaned_corpus.csv', index = False)

**Note that 'fountainpens' = 1 and 'pens' = 0**

corpus.info()

# The fountainpens subreddit make up 55.6% of the data and the pens subreddit make up 44%
corpus['subreddit'].value_counts(normalize = True) * 100
