==> code/03_nlp_processing.ipynb <==
# NLP Processing With SpaCy
---

## Contents
---
- [Data Retrieval](#Data-Retrieval)
- [SpaCy Processing](#SpaCy-Processing)

### Data Retrieval
___

**Library Imports**

import pandas as pd
import re
import spacy

**Read in cleaned_corpus.csv**

cleaned_corpus = pd.read_csv('../data/cleaned_corpus.csv')

cleaned_corpus.head()

### SpaCy Processing
---

# Load the medium size pipeline
nlp = spacy.load('en_core_web_md')

**Function that allows SpaCy to process text data so that it can be ran through an apply method for the 'text' column of the 'cleaned_corpus' dataset.**

# Function created using SpaCy lesson, references are https://spacy.io/api/token, https://realpython.com/natural-language-processing-spacy-python/#lemmatization, and ChatGPT for stucture help.  Hank reminded me that .apply will apply a function to a datafram column.

def spacy_processor(text):
    
    #Put the data into spaCy model
    doc = nlp(text)
    
    # Create a tokens list with only alpha characters and leaving out any that are only one letter - which I saw during initial tests
    # Also lemmatizes words and omits SpaCy stop words
    tokens = [token.lemma_.lower().strip() for token in doc if token.is_alpha and not token.is_stop and len(token.text) > 1]

    #Put the processed text back together
    processed_text = ' '.join(tokens)

    #return processed text to dataframe
    return processed_text
    

# Apply the function to the text column of the cleaned_corpus dataframe
cleaned_corpus['processed_text'] = cleaned_corpus['text'].apply(spacy_processor)

#Checking out how it looks
cleaned_corpus.head()

#Checking out the first entry in the processed_text column
cleaned_corpus['processed_text'][0]

# OK to drop the text column
cleaned_corpus.drop(columns = 'text', inplace = True)

cleaned_corpus.head()

I noticed when modeling that after reading in the data, there were four NaNs.  I realized that four of the rows in the dataset at thos point only had ''.  While not considered a NaN, there is no data here.  I will drop these columns.  There were either just emojis for text or stop words.

cleaned_corpus[cleaned_corpus['processed_text'] == '']


cleaned_corpus.drop([749,1140,2604,2661], inplace = True)

cleaned_corpus.shape

cleaned_corpus.to_csv('../data/text_processed_corpus.csv', index = False)
